FIXED: macOS error during authorization
This commit goes with changes in authorization scheme.

ISSUE: macOS (new releases) generates an error on startUpdatingLocation() if a user makes not immediately decision, 2 or 3 sec, with Current Location Diolog.

FIXED: Restrict error notifiying in case when a user tries to give a permission so that there is no difference in Current Location Diolog behavior in either early or new macOS releases.